requirements.txt
asgiref==3.7.2 ; python_version >= "3.9" and python_version < "4.0"
django==4.2.2 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.6 ; python_version >= "3.9" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.6.3 ; python_version >= "3.9" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"

Django~=4.2.2
djangorestframework~=3.14.0